refactor #135 : feedback history 조회 함수 feedback-history feature로 분리
